refactor(settings): move the team default agent into General, drop the rest
The team default agent is not a sharing setting — it only lived in Team Shared
because that section existed. It moves to General as `TeamDefaultAgentCard`,
restyled to the `SettingCard` idiom and placed next to Team & Organization; it
renders nothing without a current team.

Everything else the section hosted is deleted:
- `TeamOssSyncStatus` and the two panels only it rendered, `TeamSyncPaths` and
  `TeamShareDisconnect`. Disconnect could not work anyway: the cloud `DELETE
  /v1/teams/:id/share-mode` answers 410 `share_mode_permanent`.
- `TeamShareSection` and `EnableShareWizard`, which were already unreferenced on
  origin/main — enabling lives in the Knowledge column (`KnowledgeEnablePanel`).

44 more dead i18n keys removed from both locales.

Leaves `team_disconnect_repo` and `team_sync_paths` registered as Tauri commands
with no caller.
