+++ packages/app/src/locales/en.json
@@ -2153,19 +2153,8 @@
     },
     "team": {
       "oss": {
-        "title": "OSS sync",
-        "status": "Status",
-        "idle": "Idle",
-        "error": "Error",
-        "syncingLabel": "Syncing…",
-        "lastSync": "Last sync",
-        "never": "Not synced yet",
-        "pulled": "Pulled",
-        "pushed": "Pushed",
-        "conflicts": "Conflicts",
         "syncNow": "Sync now"
       },
-      "noWorkspace": "No workspace selected",
       "noCurrentTeam": "No team selected",
       "availableModelsHint": "Gateway models (click to add)",
       "hostLlm": "Host LLM (team shared AI model)",
@@ -2173,17 +2162,9 @@
       "saveLlm": "Save",
       "joinedAt": "Joined",
       "syncNow": "Sync Now",
-      "disconnect": "Disconnect",
-      "disconnectTitle": "Disconnect Team Shared Directory",
-      "disconnectConfirm": "Are you sure you want to disconnect the team shared directory? The {{teamRepoDir}} directory and all its content will be permanently deleted.",
-      "confirmDisconnect": "Disconnect",
-      "disconnectSuccess": "Local team shared directory removed",
-      "disconnectSuccessHint": "Cloud binding cleared. You can choose OSS, managed Git, or self-hosted Git again.",
       "disconnectMigrationRequired": "Cloud database is missing the disable_team_share migration. Apply services/supabase/migrations/20260604120000_disable_team_share.sql and redeploy FC.",
       "disconnectSchemaDrift": "Cloud API schema is out of date (disable_team_share). Deploy the latest FC and Supabase migration, then retry.",
       "cloudShareRequiredBeforeSync": "Team share is not locked on the cloud yet. Use Enable (Self-hosted Git) first — Sync Now cannot enable it.",
-      "desktopOnly": "Team shared directory actions require the desktop app",
-      "disconnectDesc": "This will delete local team data ({{teamcluDir}} and {{teamRepoDir}} directories). This action cannot be undone.",
       "addModel": "Add",
       "llmApiKeyDefault": ", defaults to device ID",
       "llmApiKeyHint": "API key is read from env var",
@@ -2784,35 +2765,6 @@
       "saveAndApply": "Save & Apply",
       "saving": "Applying...",
       "title": "Permission Management"
-    },
-    "teamShare": {
-      "enableTitle": "Enable team sharing",
-      "daemonDeliveryFailed": "Team share is on, but the daemon didn't accept the secret",
-      "enableDescription": "Choose a team sharing mode. It cannot be changed once enabled, so choose carefully.",
-      "modeOssDesc": "Use Alibaba Cloud OSS as the sharing backend (default)",
-      "lockWarning": "⚠️ Cannot be changed once enabled.",
-      "confirmEnable": "Confirm and enable",
-      "title": "Team sharing",
-      "lockSubtitle": "Once a team sharing mode is enabled, it cannot be changed.",
-      "signInRequired": "Please sign in to manage team sharing.",
-      "notEnabled": "Team sharing is not enabled",
-      "ownerOnlyDisconnect": "Only the team owner can disconnect team share",
-      "enable": "Enable",
-      "enabledPrefix": "Enabled:",
-      "globalSyncDirPrefix": "Global sync directory: ",
-      "workspaceLinkPrefix": "This workspace link: ",
-      "linkStatus": {
-        "linked": "Linked",
-        "pendingMigration": "Local directory (pending migration)",
-        "unlinked": "Not linked"
-      },
-      "copyPath": "Copy path",
-      "syncPaths": "Sync paths",
-      "realSyncDir": "Real sync directory",
-      "dirExists": "Exists",
-      "dirNotCreated": "Not yet created",
-      "workspaceSymlinks": "Workspace symlinks ({{count}})",
-      "current": "Current"
     },
     "teamSecret": {
       "storeUnreadable": "Cannot read the local secret store: {{msg}}",
