+++ packages/app/src/locales/zh-CN.json
@@ -2153,19 +2153,8 @@
     },
     "team": {
       "oss": {
-        "title": "OSS 同步",
-        "status": "状态",
-        "idle": "空闲",
-        "error": "错误",
-        "syncingLabel": "同步中…",
-        "lastSync": "最后同步",
-        "never": "尚未同步",
-        "pulled": "已拉取",
-        "pushed": "已推送",
-        "conflicts": "冲突",
         "syncNow": "立即同步"
       },
-      "noWorkspace": "未选择 workspace",
       "noCurrentTeam": "当前没有团队",
       "availableModelsHint": "网关可用模型（点击添加）",
       "hostLlm": "托管 LLM（团队共享 AI 模型）",
@@ -2173,17 +2162,9 @@
       "saveLlm": "保存",
       "joinedAt": "加入时间",
       "syncNow": "立即同步",
-      "disconnect": "断开连接",
-      "disconnectTitle": "断开团队共享目录",
-      "disconnectConfirm": "确定要断开团队共享目录吗？{{teamRepoDir}} 目录及其所有内容将被永久删除。",
-      "confirmDisconnect": "断开连接",
-      "disconnectSuccess": "已删除本机团队共享目录",
-      "disconnectSuccessHint": "已清除云端绑定，可重新选择 OSS / 托管 Git / 自建 Git。",
       "disconnectMigrationRequired": "云端数据库缺少 disable_team_share 迁移。请应用 services/supabase/migrations/20260604120000_disable_team_share.sql 并重新部署 FC。",
       "disconnectSchemaDrift": "Cloud API 架构过旧（缺少 disable_team_share）。请部署最新 FC 与 Supabase 迁移后重试。",
       "cloudShareRequiredBeforeSync": "云端尚未锁定团队共享。请先使用「开通」（自托管 Git），「立即同步」无法代替开通。",
-      "desktopOnly": "团队共享目录操作需在桌面客户端中进行",
-      "disconnectDesc": "此操作将删除本地团队数据（{{teamcluDir}} 和 {{teamRepoDir}} 目录），无法撤销。",
       "addModel": "添加",
       "llmApiKeyDefault": "。",
       "llmApiKeyHint": "API 密钥从环境变量读取",
@@ -2784,35 +2765,6 @@
       "saveAndApply": "保存并应用",
       "saving": "应用中...",
       "title": "权限管理"
-    },
-    "teamShare": {
-      "enableTitle": "开通团队共享",
-      "daemonDeliveryFailed": "团队共享已开启，但 daemon 未接收密钥",
-      "enableDescription": "选择团队共享模式。开通后不可切换，请谨慎选择。",
-      "modeOssDesc": "使用阿里云 OSS 作为共享后端（默认）",
-      "lockWarning": "⚠️ 开通后不可切换。",
-      "confirmEnable": "确认开通",
-      "title": "团队共享",
-      "lockSubtitle": "团队共享模式一经开通，不可切换。",
-      "signInRequired": "请先登录后再管理团队共享。",
-      "notEnabled": "团队共享未开通",
-      "ownerOnlyDisconnect": "仅团队 owner 可断开团队共享",
-      "enable": "开通",
-      "enabledPrefix": "已开通:",
-      "globalSyncDirPrefix": "全局同步目录：",
-      "workspaceLinkPrefix": "本工作区链接：",
-      "linkStatus": {
-        "linked": "已软链",
-        "pendingMigration": "本地目录（待迁移）",
-        "unlinked": "未链接"
-      },
-      "copyPath": "复制路径",
-      "syncPaths": "同步路径",
-      "realSyncDir": "真实同步目录",
-      "dirExists": "存在",
-      "dirNotCreated": "尚未创建",
-      "workspaceSymlinks": "工作区软链（{{count}}）",
-      "current": "当前"
     },
     "teamSecret": {
       "storeUnreadable": "无法读取本机密钥库：{{msg}}",
